添加完成客户信息表的分页展示和服务器排序，添加easyui组件

+++ hzitcrm-dao/src/main/java/com/hzit/crm/core/mapper/CustomerInfoMapper.java
@@ -30,5 +30,5 @@ public interface CustomerInfoMapper {
 	 * 获取客户表的总记录数
 	 * @return
      */
-	public int getTotal();
+	public long getTotal();
 }
